separate settings and enums
